chore: Moved `<AccordionSection />` and `<DimensionedChart />` to composed components
