+++ front-end-components/index.html
@@ -92,14 +92,14 @@
       ></div>
       <button type="submit">Submit</button>
     </form>-->
-    <!--<div id="compare-your-trust" data-id="09617166"></div>-->
+    <div id="compare-your-trust" data-id="00000001"></div>
     <!--<div
       id="compare-your-costs"
       data-type="trust"
       data-id="02387916"
       data-phases='["Secondary","Primary","Post-16"]'
     ></div>-->
-    <div id="compare-your-costs" data-type="school" data-id="990095"></div>
+    <!--<div id="compare-your-costs" data-type="school" data-id="990095"></div>-->
     <!--<div id="historic-data" data-type="school" data-id="990095"></div>-->
     <!--<div id="historic-data-2" data-type="school" data-id="123456" data-phase="Primary" data-finance-type="Maintained"></div>-->
     <!--<div
