+++ front-end-components/src/composed/accordion-section/index.tsx
@@ -1,0 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from "src/composed/accordion-section/types";
+export * from "src/composed/accordion-section/composed";
