+++ front-end-components/src/composed/accordion-section/types.tsx
@@ -1,0 +1,12 @@
+import {
+  SchoolChartData,
+  TrustChartData,
+} from "src/components/charts/table-chart";
+import { DimensionedChartProps } from "../dimensioned-chart";
+
+export type AccordionSectionProps<
+  TData extends SchoolChartData | TrustChartData,
+> = Omit<DimensionedChartProps<TData>, "dimensions" | "topLevel"> & {
+  index: number;
+  title: string;
+};
