+++ front-end-components/src/composed/dimensioned-chart/index.tsx
@@ -1,0 +1,3 @@
+/* eslint-disable react-refresh/only-export-components */
+export * from "src/composed/dimensioned-chart/types";
+export * from "src/composed/dimensioned-chart/composed";
