+++ front-end-components/src/views/compare-your-costs/partials/accordion-sections/educational-ict.tsx
@@ -13,7 +13,7 @@ import { CostCategories, PoundsPerPupil } from "src/components";
 import { PhaseContext, CustomDataContext } from "src/contexts";
 import { HorizontalBarChartWrapperData } from "src/composed/horizontal-bar-chart-wrapper";
 import { ExpenditureApi, EducationalIctExpenditure } from "src/services";
-import { AccordionSection } from "./accordion-section";
+import { AccordionSection } from "src/composed/accordion-section";
 
 export const EducationalIct: React.FC<CompareYourCostsProps> = ({
   type,
