+++ front-end-components/src/views/compare-your-costs/partials/accordion-sections/educational-supplies.tsx
@@ -13,7 +13,7 @@ import { CostCategories, PoundsPerPupil } from "src/components";
 import { CustomDataContext, PhaseContext } from "src/contexts";
 import { HorizontalBarChartWrapperData } from "src/composed/horizontal-bar-chart-wrapper";
 import { ExpenditureApi, EducationalSuppliesExpenditure } from "src/services";
-import { AccordionSection } from "./accordion-section";
+import { AccordionSection } from "src/composed/accordion-section";
 
 export const EducationalSupplies: React.FC<CompareYourCostsProps> = ({
   type,
