+++ front-end-components/src/views/compare-your-costs/partials/accordion-sections/index.tsx
@@ -8,4 +8,3 @@ export * from "src/views/compare-your-costs/partials/accordion-sections/educatio
 export * from "src/views/compare-your-costs/partials/accordion-sections/educational-ict";
 export * from "src/views/compare-your-costs/partials/accordion-sections/catering-staff-services";
 export * from "src/views/compare-your-costs/partials/accordion-sections/administrative-supplies";
-export * from "src/views/compare-your-costs/partials/accordion-sections/accordion-section";
