+++ front-end-components/src/views/compare-your-costs/partials/accordion-sections/utilities.tsx
@@ -13,7 +13,7 @@ import { PoundsPerMetreSq, PremisesCategories } from "src/components";
 import { PhaseContext, CustomDataContext } from "src/contexts";
 import { HorizontalBarChartWrapperData } from "src/composed/horizontal-bar-chart-wrapper";
 import { ExpenditureApi, UtilitiesExpenditure } from "src/services";
-import { AccordionSection } from "./accordion-section";
+import { AccordionSection } from "src/composed/accordion-section";
 
 export const Utilities: React.FC<CompareYourCostsProps> = ({ type, id }) => {
   const [dimension, setDimension] = useState(PoundsPerMetreSq);
